3D rendering with uniform buffers, refactor Win32/Vulkan utils
Slop overview:
Introduce uniform buffer support and refactor platform/Vulkan utilities.

- Implemented uniform buffer object (vkUniformBufferData) and full descriptor set/layout/pool management in vkMeshClass (initUniformBuffers, initDescriptorSetLayout, initDescriptorPoolAndSets, updateUniformBuffer, draw, cleanup).

- Updated vertex shader to read from a UBO and generate SPIR-V asset.

- Replaced vkFunctionPtrs.h with vkExtFunctions implementation in vkUtils; moved debug utils function pointers into vkUtils.cpp.

- Refactored vkRenderingContext to integrate mesh UBOs (initDescriptorSetLayout, using mesh descriptor set layout in pipeline layout, call updateUniformBuffer before draw, adjust front-face winding), and improved swapchain recreation logic.

- Renamed and standardized Win32 utility types/functions (win32_utils -> win32Utils) and updated appRuntime/window event APIs (windowEvent_* methods, close->forceClose) and related call sites.

- Fixed format specifiers in memoryUtils to use size_t (%zu), replaced clamp calls with std::clamp and added required includes.

- Added minor app-level features: mesh rotation state and propagation (appInterface, vkRenderingContext, vkMeshClass), debug helper in utilFunctions, and various cleanups/clarifications in log messages and comments.

This change prepares the codebase for per-frame uniform updates, descriptor management, and tidies platform/Vulkan helper functions. There is still more work to be done, as always, but I'll leave it at that for today.

+++ ibex3D_sampleApplication/include/ibex3D/utility/memoryUtils.h
@@ -13,7 +13,7 @@ static memoryStats_t memoryStats;
 void* operator new(size_t size)
 {
 	memoryStats.totalUsage += size;
-	printf("Allocated %d bytes (now using %d bytes).\n", (int)size, (int)memoryStats.totalUsage);
+	printf("Allocated %zu bytes (now using %zu bytes).\n", size, memoryStats.totalUsage);
 
 	return malloc(size);
 }
@@ -21,7 +21,7 @@ void* operator new(size_t size)
 void operator delete(void* memory, size_t size)
 {
 	memoryStats.totalUsage -= size;
-	printf("Deallocated %d bytes (now using %d bytes).\n", (int)size, (int)memoryStats.totalUsage);
+	printf("Deallocated %zu bytes (now using %zu bytes).\n", size, memoryStats.totalUsage);
 
 	free(memory);
 }
